backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
requests==2.32.3
python-dotenv==1.0.1
tenacity==9.0.0
google-genai==0.7.0


